From a678d9087b05aaf6947e15f0cc9a93ace299466f Mon Sep 17 00:00:00 2001 Subject: [PATCH] Fix ifinfo_link_with_port race condition with newlink
The race condition could happen like this:
When an interface is enslaved into the port channel immediately after
it is created, the order of creating the ifinfo and linking the ifinfo to
the port is not guaranteed.

The team handler will listen to both netlink message to track new links
get created to allocate the ifinfo and add the ifinfo into its linked list,
and the team port change message to link the new port with ifinfo found
in its linkedin list. However, when the ifinfo is not yet created, the error
message "Failed to link port with ifinfo" is thrown with member port failed
to be added into the team handler's port list.

This fix adds a condition to check if ifinfo_link_with_port is linking ifinfo
to a port or to the team interface itself. If it is a port, ifinfo_find_create
function is used to fix the race condition.